Wrap DiffCalculator
