+++ station/Classes/Core/DiffCalculator/Impl/Models/CellChanges.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+class CellChanges {
+    var inserts = [IndexPath]()
+    var deletes = [IndexPath]()
+    var reloads = [IndexPath]()
+
+    init(inserts: [IndexPath] = [], deletes: [IndexPath] = [], reloads: [IndexPath] = []) {
+        self.inserts = inserts
+        self.deletes = deletes
+        self.reloads = reloads
+    }
+}
