+++ station/Classes/Core/DiffCalculator/Impl/Models/ReloadableCell.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+struct ReloadableCell<N: Equatable>: Equatable {
+    var key: String
+    var value: N
+    var index: Int
+
+    static func == (lhs: ReloadableCell, rhs: ReloadableCell) -> Bool {
+        return lhs.key == rhs.key && lhs.value == rhs.value
+    }
+}
